new part
Added new part of the website. Only displays “Coming Soon“, but is
going to be the about section. Still debating on if I want to get rid
of the home link in the nav bar, probably a good idea, but what am I
going to put in its place (if anything).