I'm back... and I've gotten a lot done (7/18/17)
I’ve got a basic design. I’m working with the black bar design with its
color scheme. I set up the parallax parts, so all I have to do is the
rest of the page. I put a “coming soon” animation on the incomplete
parts and finished the contact me section. This update is live.