It's beginning to look a lot like Christmas! (7/19/17)
Dumb title, shut up. This is a big update and I thought that sounded
funny since I couldn't think of anything better.

Anyway, started working on the about section (for like the billionth
time, but this time it’s for real, I promise… maybe). Added a “coming
soon” header to each section that’s not complete and got rid of the cool
coming soon animation (because redundant headers are redundant, yo).
Also added an incomplete, but working/responsive footer to the about
page. Stylistically it’s finished, but in terms of content, it’s not
done. The first box is done, but the last 3 I still need to sort out. I
also changed the parallax images to make the website look like it has
more of a constant color scheme. I still need to get it to look right
on mobile devices, but I can do that later. Most importantly, I made
the home section more responsive. The logo and title resize to the
height of the page and works on every screen.

*This update is live (buildtect.com)*